requirements.txt
Django==2.2
django-choices==1.7.2
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
importlib-metadata==4.8.1
Markdown==3.3.4
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uuid==1.30
zipp==3.6.0
